// src/features/Alerts/MountingShow/types/mountingShowTypes.ts
interface MountingShowResponse {
  alert: Alert;
  alert_cause: string;
  alert_id: number;
  alert_vehicle_tire_id: number;
  comment: string;
  created_at: string;
  created_by: CreatedBy;
  tire_review: TireReview;
  tire_review_id: number;
  updated_at: string;
  updated_by: CreatedBy;
  vehicle_tire: VehicleTire;
  vehicle_tire_id: number;
}

interface TireReview {
  accumulated_mileage: number;
  activity: string;
  archived: number;
  comment: string;
  created_at: string;
  created_by: number;
  date: string;
  history: number;
  last_review_month: number;
  last_travel: number;
  movement_tire: MovementTire;
  movement_tire_id: number;
  old_tire_review_id: number;
  pressure: string;
  sensor_check: number;
  tire_review_depth_line: TireReviewDepthLine;
  tire_review_id: number;
  updated_at: string;
  updated_by: CreatedBy;
  vehicle_review_id: number;
}

interface MovementTire {
  archived: number;
  created_at: string;
  created_by: number;
  history: number;
  movement: string;
  movement_tire_id: number;
  old_movement_tire_id: number;
  tire_cycle: TireCycle;
  tire_cycle_id: number;
  updated_at: string;
  updated_by: number;
}
// ...
interface VehicleTire {
  created_at: string;
  created_by: CreatedBy;
  history: number;
  movement_tire_id: number;
  updated_at: string;
  updated_by: CreatedBy;
  vehicle: Vehicle;
  vehicle_id: number;
  vehicle_tire_id: number;
  vehicle_type_axle_tire_id: number;
}

interface Vehicle {
  archived: number;
  created_at: string;
  created_by: number;
  cylinder_capacity: number; // TODO: check if this is correct
  economic_number: string;
  engine_transmission_brand_id: number; // TODO: check if this is correct
  enrollment: string;
  has_odometer: number;
  old_vehicle_id: number;
  status: number;
  subsidiary: Subsidiary;
  subsidiary_id: number;
  transmission_model: string; // TODO: check if this is correct
  transmission_speeds: number; // TODO: check if this is correct
  type_of_route: string;
  updated_at: string;
  updated_by: number; // TODO: check if this is correct
  vehicle_brand: VehicleBrand;
  vehicle_brand_id: number;
  vehicle_driver: number; // TODO: check if this is correct
  vehicle_engine_brand_id: number; // TODO: check if this is correct
  vehicle_id: number;
  vehicle_model: string; // TODO: check if this is correct
  vehicle_type: VehicleType;
  vehicle_type_id: number;
  vehicle_year: number; // TODO: check if this is correct
  wheel: string;
}